refactor: extract internal packages for validation and binding
  Move implementation details to internal/validator and internal/bind
  to reduce public API surface. Keep root package focused on public
  exports and integration. Fixes magic number issue in rate limiting.
